REMOVE of feature: randomnum, calculator
This commmit should be historical, as I deleted calculator.
I copied these code from CSDN in my primary school,
with one of my best friends when I am in grade 6.
That day I secretly brought my iPod Touch 6 to the school,
without telling the teacher.
This experience should be remembered forever, as it is, and will always
be a wonderful memory to me.
Also, I used neither git nor other VCS back then,
only a NetDisk to do the job, so this history was not recorded before.